managementPoint/requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distlib==0.3.8
Django==5.0.4
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
Flask==3.0.0
idna==3.7
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.6
MarkupSafe==2.1.3
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
platformdirs==4.1.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
Werkzeug==3.0.1